# Fetch records concurrently with benchmarks