--- a/pkg/blockbuilder/blockbuilder.go
+++ b/pkg/blockbuilder/blockbuilder.go
@@ -586,6 +586,8 @@ func (b *BlockBuilder) consumePartition(
 				loopDone = true
 			}
 		})
+
+		b.kafkaClient.AllowRebalance()
 	}
 
 	close(recsC)