A quick change to see if we make fetching records and processing them concurrent, do we see any gains. I wrote a quick benchmark, and the numbers vary greatly with different number of records/samples/series. With the benchmark as is, I see ~16% gains.

Although https://github.com/grafana/mimir/pull/8711 is underway, it will be nice to try this out in the meantime.

```
benchmark                    old ns/op      new ns/op      delta
BenchmarkBlockBuilder-10     3157905625     2640359666     -16.39%
```