@codesome 2024-07-16
I just fixed a race and I see more gains. 16%. Updated PR description.