@narqo 2024-07-16
_We've discussed that in the DM, but I'm putting it here, just so it didn't slip, and I didn't forget while I'm off._

The idea looks good to me. But, in its current implementation, I'm concerned this re-introduces the issue with Kafka rebalance, that we were addressing in #8591. That is, a consumer that is a part of a group, should not block the group's rebalance for extended period of time. Otherwise, after around one minute of blockage (or so), the broker kicks the consumer out of the group, assuming it is malfunctioning.