@codesome 2024-07-16
Updated the PR to call `AllowRebalance` more often. Making this PR work same as before without only fetching and processing made concurrent. I will test it out and see how it goes.