@codesome 2024-07-17
I tested this, and it ended up taking a little longer to consume and the memory consumption was much higher (3x). Memory usage might be because of a high channel size, leading to a lot of records buffering up. But the time increase is unexpected.

After rolling back this change in dev, the memory consumption stayed higher with consumption time still staying higher. So I am not sure if this test was conclusive.

Need to test it again in a controlled setting where the load does not change much between tests.